feat: Added CharacterLocation entities

+++ RickAndMorty/RickAndMortyTests/Entities/DBEntities/DBCharacterUnitTests.swift
@@ -20,7 +20,7 @@ final class DBCharacterUnitTests: XCTestCase {
         let character = dbCharacter.convertToEntity()
         
         // Then
-        XCTAssertEqual(Int32(character.id), dbCharacter.id)
+        XCTAssertEqual(Int16(character.id), dbCharacter.id)
         XCTAssertEqual(character.name, dbCharacter.name)
         XCTAssertEqual(character.status.rawValue, dbCharacter.status)
         XCTAssertEqual(character.species, dbCharacter.species)
